pyproject.toml
[tool.poetry]
name = "crypto-data-provider"
version = "0.1.0"
description = ""
authors = ["Aleksei Gruzdev <algruwork@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.103.1"
uvicorn = "^0.23.2"
pydantic-settings = "^2.0.3"
pytz = "^2023.3.post1"
python-okx = "^0.2.1"
pyyaml = "^6.0.1"
sqlalchemy = "^2.0.20"
alembic = "^1.12.0"
psycopg2-binary = "^2.9.7"


[tool.poetry.group.test.dependencies]
pytest = "^7.4.1"
httpx = "^0.24.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
